requirements.txt
streamlit==1.12.0
Pillow==9.2.0
phi-agent==0.1.0
python-dotenv==0.19.1
groq
firebase-admin
